# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenDeX Backend API",
    "version": "0.1.0",
    "summary": "Local control API of the OpenDeX backend (Android multi-window mirroring).",
...
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "contact": {
      "name": "OpenDeX on GitHub",
      "url": "https://github.com/Lordof21/OpenDex"
    }
  },
  "paths": {
...
    "/api/v1/audio/apps": {
      "get": {
        "tags": [
          "audio"
        ],
        "summary": "List App Audio",
        "description": "{supported, mode, apps, sync}: `supported` is false on Android ≤12 / an old daemon jar (single legacy stream).\n`sync` = {supported, offset_ms, pc_output_ms, link_ms, target_ms, late_extra_ms}: the \"İkisi\" (DeX + phone) alignment.",
        "operationId": "list_app_audio_api_v1_audio_apps_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List App Audio Api V1 Audio Apps Get"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          }
        }
      }
    },
    "/api/v1/audio/sync": {
      "put": {
        "tags": [
          "audio"
        ],
        "summary": "Put Audio Sync",
        "description": "The DeX page reports what it measures about its own audio: the output device's latency and how many chunks of the\nlast window were too late for the common target. The phone and the page then move to a new common target (see\nstreams/app_audio.py). The fine tune is a setting (`audio_sync_offset_ms` in PUT /api/settings).",
        "operationId": "put_audio_sync_api_v1_audio_sync_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioSyncReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Put Audio Sync Api V1 Audio Sync Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/audio/clock": {
      "post": {
        "tags": [
          "audio"
        ],
        "summary": "Audio Clock",
        "description": "The device's monotonic clock (µs) — the clock of the audio PTS. The page calls this a few times, times each round\ntrip with its own clock and keeps the quickest, to know which PTS is \"now\" on its side. 503 when the daemon cannot say.",
        "operationId": "audio_clock_api_v1_audio_clock_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Audio Clock Api V1 Audio Clock Post"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/audio/probe": {
      "post": {
        "tags": [
          "audio"
        ],
        "summary": "Audio Probe",
        "description": "The phone's half of the \"İkisi\" calibration: test tones on the phone at the instants (device-clock PTS) in the answer.\nThe page plays its own tone at the matching instants, listens to both with the laptop's microphone and writes the\nfine tune (`audio_sync_offset_ms`) from the gap it hears. 409 `not_supported` when this device cannot align.",
        "operationId": "audio_probe_api_v1_audio_probe_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Audio Probe Api V1 Audio Probe Post"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/audio/apps/{package}": {
      "put": {
        "tags": [
          "audio"
        ],
        "summary": "Put App Audio",
        "description": "Saves the package's preference (device-independent) and applies it to its open windows, if any. With\n`standalone: true` (the Media Center's transfer) an app WITHOUT a window is played on the PC too; the answer carries\n`error` when the phone could not carry it out (the transfer is then already taken back). 409 + a code\n(`not_supported`: no per-app audio on this device; `internal_package`) when it cannot start at all.",
        "operationId": "put_app_audio_api_v1_audio_apps__package__put",
        "parameters": [
          {
            "name": "package",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255,
              "pattern": "^[A-Za-z_][A-Za-z0-9_]*(\\.[A-Za-z_][A-Za-z0-9_]*)*$",
              "title": "Package"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppAudioPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Put App Audio Api V1 Audio Apps  Package  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppAudioPatch": {
        "properties": {
          "route": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "pc",
                  "both",
                  "phone"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Route"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "muted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Muted"
          },
          "standalone": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standalone"
          }
        },
        "type": "object",
        "title": "AppAudioPatch",
        "description": "PUT /api/audio/apps/{package}: only the fields present change."
      },
...
      "AudioSyncReport": {
        "properties": {
          "pc_output_ms": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pc Output Ms"
          },
          "late_chunks": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Late Chunks",
            "default": 0
          }
        },
        "type": "object",
        "title": "AudioSyncReport",
        "description": "PUT /api/audio/sync: what the DeX page measures about its own audio — both optional, the page sends what changed."
      },
...
      "ErrorDetail": {
        "title": "ErrorDetail",
        "type": "object",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "array"
              },
              {
                "type": "object"
              }
            ],
            "description": "A human-readable message, a machine code (e.g. `not_supported`), or — for 422 — the list of validation errors."
          }
        },
        "required": [
          "detail"
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The per-user API token (`OPENDEX_API_TOKEN` or `~/.opendex/api-token`)."
      },
      "tokenQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "The same token as a query parameter — accepted on GET requests and WebSocket handshakes only."
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "The API token is missing or wrong (WebSocket: close code 4401).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "ForeignOrigin": {
        "description": "The request carries an `Origin` that is not one of the app's own origins, or a `Host` that is not a loopback name (WebSocket: close code 4403).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "LengthRequired": {
        "description": "The request body was sent with chunked transfer-encoding and no `Content-Length` (chunked bodies are refused).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "The request body is larger than `OPENDEX_MAX_BODY_BYTES` (default 2 MB).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8710",
      "description": "Default local listener (OPENDEX_HTTP_HOST / OPENDEX_HTTP_PORT)"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "tokenQuery": []
    }
  ]
}